Remove extra words in show name during searches
These words doesn't belong to the original show name, but there they
are! For example, "House of Cards" appears on TheTVDB.com as "House of
Cards (US)" or "House of Cards (2013)" and, therefore searches on
BJShare may return no data when using anything but the original show
name.

Funny is that SickRage needs these words to validate the search results.
So, we cannot simply strip them out. Hence, we remove them temporally
during the search and reattach later.